onboarding-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pgvector>=0.2.0
sqlalchemy>=2.0.35
alembic>=1.14.0
psycopg2-binary>=2.9.5
pydantic>=2.7.4
python-multipart>=0.0.12
beautifulsoup4>=4.13.4
scrapy>=2.13.3
requests>=2.32.5
aiofiles>=24.1.0
pypdf2>=3.0.1
python-docx>=1.2.0
docx2txt>=0.8
python-dotenv>=1.1.1
celery>=5.5.3
redis>=5.2.0
minio>=7.2.16
openai>=1.101.0
jinja2>=3.1.0
loguru>=0.7.0
structlog>=24.0.0
pillow>=11.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4